Fix ambiguous symbol issues
