+++ mhw-cs-plugin-loader/TextureManager.cpp
@@ -1,20 +1,15 @@
 #include "TextureManager.h"
 #include "HResultHandler.h"
 
-#include <filesystem>
-
-#include <directxtk12/ResourceUploadBatch.h>
-#include <directxtk12/DDSTextureLoader.h>
-#include <directxtk12/WICTextureLoader.h>
-#include <directxtk/DDSTextureLoader.h>
-#include <directxtk/WICTextureLoader.h>
-
-
 TextureManager::TextureManager(ID3D12Device* device, ID3D12CommandQueue* cmd_queue)
-    : m_is_d3d12(true), m_device12(device), m_command_queue12(cmd_queue) { }
+    : m_is_d3d12(true), m_device12(device), m_command_queue12(cmd_queue) {}
 
 TextureManager::TextureManager(ID3D11Device* device, ID3D11DeviceContext* context) 
-    : m_is_d3d12(false), m_device11(device), m_context11(context) { }
+    : m_is_d3d12(false), m_device11(device), m_context11(context) {}
+
+TextureManager::~TextureManager() {
+    CoUninitialize();
+}
 
 TextureHandle TextureManager::load_texture(std::string_view path) {
     TextureHandle handle;
@@ -59,95 +54,6 @@ void TextureManager::unload_texture(TextureHandle handle) {
     m_textures.erase(it);
 
     dlog::debug("Unloaded texture: handle {}", handle);
-}
-
-TextureManager::ComPtr<ID3D11ShaderResourceView> TextureManager::load_texture11(std::string_view path) const {
-    namespace fs = std::filesystem;
-
-    const auto file = fs::path(path);
-    if (!fs::exists(file)) {
-        dlog::error("Failed to load texture: {} does not exist", path);
-        return nullptr;
-    }
-
-    const auto ext = file.extension().string();
-    if (ext == ".dds") {
-        ComPtr<ID3D11Resource> texture;
-        ComPtr<ID3D11ShaderResourceView> srv;
-        HandleResult(DirectX::CreateDDSTextureFromFile(
-            m_device11,
-            file.c_str(),
-            texture.GetAddressOf(),
-            srv.GetAddressOf()
-        ));
-
-        return srv;
-    }
-
-    if (ext == ".png" || ext == ".jpg" || ext == ".jpeg") {
-        ComPtr<ID3D11Resource> texture;
-        ComPtr<ID3D11ShaderResourceView> srv;
-
-        Microsoft::WRL::Wrappers::RoInitializeWrapper initialize(RO_INIT_MULTITHREADED);
-        HandleResult(DirectX::CreateWICTextureFromFile(
-            m_device11,
-            m_context11,
-            file.c_str(),
-            texture.GetAddressOf(),
-            srv.GetAddressOf()
-        ));
-
-        return srv;
-    }
-
-    dlog::error("Failed to load texture: unsupported format {}", ext);
-    return nullptr;
-}
-
-TextureManager::ComPtr<ID3D12Resource> TextureManager::load_texture12(std::string_view path) const {
-    namespace fs = std::filesystem;
-
-    const auto file = fs::path(path);
-    if (!fs::exists(file)) {
-        dlog::error("Failed to load texture: {} does not exist", path);
-        return nullptr;
-    }
-
-    const auto ext = file.extension().string();
-    if (ext == ".dds") {
-        ComPtr<ID3D12Resource> texture;
-        DirectX::ResourceUploadBatch resource_upload(m_device12);
-
-        HandleResult(DirectX::CreateDDSTextureFromFile(
-            m_device12,
-            resource_upload,
-            file.c_str(),
-            texture.GetAddressOf()
-        ));
-
-        resource_upload.End(m_command_queue12).wait();
-
-        return texture;
-    }
-
-    if (ext == ".png" || ext == ".jpg" || ext == ".jpeg") {
-        ComPtr<ID3D12Resource> texture;
-        DirectX::ResourceUploadBatch resource_upload(m_device12);
-
-        HandleResult(DirectX::CreateWICTextureFromFile(
-            m_device12,
-            resource_upload,
-            file.c_str(),
-            texture.GetAddressOf()
-        ));
-
-        resource_upload.End(m_command_queue12).wait();
-
-        return texture;
-    }
-
-    dlog::error("Failed to load texture: unsupported format {}", ext);
-    return nullptr;
 }
 
 D3D12_GPU_DESCRIPTOR_HANDLE TextureManager::get_gpu_descriptor_handle(TextureEntry& entry) {
